timeline folder restructure
