+++ packages/timeline/src/services/index.ts
@@ -1,0 +1,11 @@
+// Main timeline service (singleton)
+export { default as timelineService } from './timeline/timeline.service';
+export { TimelineService } from './timeline/timeline.service';
+
+// Data management services
+export { TimelineDataService } from './timeline/timeline-data.service';
+
+// Feature services
+export { ElementService } from './element/element.service';
+export { AnimationService } from './animation/animation.service';
+export { CaptionService } from './caption/caption.service';
