+++ packages/timeline/src/types.ts
